⚗️  Move to modular Slate editor
