+++ apps/client/components/Editable/plugins/plugins.interface.ts
@@ -1,0 +1,9 @@
+import { Editor } from 'slate';
+
+export type EditablePlugin = {
+  wrapper: (editor: Editor) => Editor;
+  onKeyDown?: (
+    editor: Editor,
+    event: React.KeyboardEvent<HTMLDivElement>
+  ) => void;
+};
